[kernel] Build everything with -Wconversion [part 36]
The kernel static no-arch target builds with clang + WCONV in 64-bit mode
for the unit tests, but the tests cannot be built with WCONV yet.

Changes to be committed:
	modified:   .vscode/tasks.json
	modified:   include/tilck/kernel/kmalloc.h
	modified:   include/tilck/kernel/paging.h
	modified:   include/tilck/kernel/process.h
	modified:   include/tilck/kernel/term.h
	modified:   kernel/arch/i386/paging.c
	modified:   kernel/char/tty/term_action_wrappers.c.h
	modified:   kernel/char/tty/tty_input.c
	modified:   kernel/char/tty/tty_ioctl.c
	modified:   kernel/cwd.c
	modified:   kernel/elf.c
	modified:   kernel/fs/devfs.c
	modified:   kernel/fs/fat32.c
	modified:   kernel/fs/fs_syscalls.c
	modified:   kernel/fs/mp.c
	modified:   kernel/fs/vfs.c
	modified:   kernel/fs/vfs_util.c
	modified:   kernel/gcov.c
	modified:   kernel/misc.c
	modified:   kernel/mm/kmalloc.c
	modified:   kernel/mm/kmalloc_heaps_debug.c.h
	modified:   kernel/mm/process_mm.c
	modified:   kernel/mm/system_mmap.c
	modified:   kernel/printk.c
	modified:   unittests/kmalloc_mocks.cpp
	modified:   unittests/kmalloc_test.cpp

+++ include/tilck/kernel/paging.h
@@ -38,27 +38,27 @@ map_page_int(page_directory_t *pdir, void *vaddr, uptr paddr, u32 flags);
 NODISCARD int
 map_zero_page(page_directory_t *pdir, void *vaddrp, bool us, bool rw);
 
-NODISCARD int
+NODISCARD size_t
 map_pages(page_directory_t *pdir,
           void *vaddr,
           uptr paddr,
-          int page_count,
+          size_t page_count,
           bool big_pages_allowed,
           bool us,
           bool rw);
 
-NODISCARD int
+NODISCARD size_t
 map_pages_int(page_directory_t *pdir,
               void *vaddr,
               uptr paddr,
-              int page_count,
+              size_t page_count,
               bool big_pages_allowed,
               u32 flags);
 
-NODISCARD int
+NODISCARD size_t
 map_zero_pages(page_directory_t *pdir,
                void *vaddrp,
-               int page_count,
+               size_t page_count,
                bool us,
                bool rw);
 
@@ -75,10 +75,10 @@ void set_page_rw(page_directory_t *pdir, void *vaddr, bool rw);
 static inline void
 unmap_pages(page_directory_t *pdir,
             void *vaddr,
-            int page_count,
+            size_t page_count,
             bool free_pageframes)
 {
-   for (int i = 0; i < page_count; i++) {
+   for (size_t i = 0; i < page_count; i++) {
       unmap_page(pdir, (char *)vaddr + (i << PAGE_SHIFT), free_pageframes);
    }
 }
