+++ include/tilck/kernel/process.h
@@ -160,7 +160,7 @@ static ALWAYS_INLINE bool is_kernel_thread(task_info *ti)
 static ALWAYS_INLINE int thread_ti_to_tid(task_info *ti)
 {
    ASSERT(ti->tid != ti->pid);
-   return MAX_PID + (sptr) ((uptr)ti - KERNEL_BASE_VA);
+   return (int)(MAX_PID + (sptr) ((uptr)ti - KERNEL_BASE_VA));
 }
 
 static ALWAYS_INLINE task_info *thread_tid_to_ti(int tid)
